Added delayed tests for im mem and redis cache managers
