+++ packages/core/test/staticClientCachingDriver.test.ts
@@ -10,6 +10,9 @@ import chaiAsPromised from "chai-as-promised";
 
 import cacheTests from "./cache/basic.tests";
 import etagTests from "./cache/etag.tests";
+import delayedTests from "./cache/delayedTests.tests";
+
+import redisEtagTests from "./cache/redisEtag.tests";
 import evictionTests from "./cache/eviction.tests";
 import multipleHeadersTests from "./cache/multipleHeaders.tests";
 import noCacheHeaderTests from "./cache/noCacheHeader.tests";
@@ -24,9 +27,10 @@ describe("Default cache tests", function() {
       baseUri: "https://somewhere"
     });
   });
-  cacheTests();
   etagTests();
+  cacheTests();
   evictionTests();
   multipleHeadersTests();
   noCacheHeaderTests();
+  delayedTests();
 });
