+++ packages/core/test_integration/staticClientCachingDriver.test.ts
@@ -8,6 +8,10 @@
 import chai from "chai";
 import chaiAsPromised from "chai-as-promised";
 
+import delayedTests from "../test/cache/delayedTests.tests";
+
+import redisEtagTests from "../test/cache/redisEtag.tests";
+
 import cacheTests from "../test/cache/basic.tests";
 import etagTests from "../test/cache/etag.tests";
 import evictionTests from "../test/cache/eviction.tests";
@@ -32,5 +36,6 @@ describe("Redis cache tests", function() {
   after(function() {
     this.client.clientConfig.cacheManager.quit();
   });
-  cacheTests();
+  etagTests();
+  delayedTests();
 });
